Fixed enchantment registration.
Enchantments should always be registered through GameRegistry, as there
are several bugs if you do not. The biggest being compat issues with
other mods, however the most notable issue is that getting the id based
on the existing enchantments will cause the enchantment ids to shift
around every time a new enchantment is added to the instance and are
thus unreliable.